+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Change _ProductName_, added: showBrandName, showEan, ean attributes.
+- Change _ProductName_, added: showBrandName, showProductReference, productReference attributes.
 
 ## [1.3.1] - 2018-5-30
 ### Fixed